LayoutTests/http/tests/media/video-play-stall-seek.html

<video controls></video>
<p>Test that playback can be resumed by seeking backwards after load stalls.</p>
<script src=../../../media/video-test.js></script>
<script>

    waitForEvent('waiting', function () {

        // now that playback has paused to wait for data, seek back and verify that we
        // get a 'canplay' event
        waitForEvent('canplay' , function () {
            testExpected("video.readyState", HTMLMediaElement.HAVE_CURRENT_DATA, ">=");
            testExpected("video.networkState", HTMLMediaElement.NETWORK_LOADING);
            endTest();
        } );
        testExpected("video.readyState", HTMLMediaElement.HAVE_CURRENT_DATA);
        
        run("video.currentTime = 0.1");
    } );

    video.src = "http://127.0.0.1:8000/media/video-load-and-stall.cgi?name=../../../media/content/test.mp4&stallAt=100000";
    run("video.play()");
</script>
