LayoutTests/http/tests/media/video-play-stall-seek-expected.txt

Test that playback can be resumed by seeking backwards after load stalls.

RUN(video.play())
EVENT(waiting)
EXPECTED (video.readyState == '2') OK
RUN(video.currentTime = 0.1)
EVENT(canplay)
EXPECTED (video.readyState >= '2') OK
EXPECTED (video.networkState == '2') OK
END OF TEST
